+++ apps/pioneer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @internal/playground
 
+## 0.4.3
+
+### Patch Changes
+
+- bump
+
+## 0.4.2
+
+### Patch Changes
+
+- bump for maya
+
 ## 0.4.1
 
 ### Patch Changes