+++ e2e/wallets/integration-wallet-package-keepkey/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @coinmasters/integration-wallet-keepkey
 
+## 0.2.3
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@10.0.3
+  - @coinmasters/types@4.4.3
+  - @coinmasters/core@8.0.3
+
+## 0.2.2
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@10.0.2
+  - @coinmasters/types@4.4.2
+  - @coinmasters/core@8.0.2
+
 ## 0.2.1
 
 ### Patch Changes