+++ packages/swapkit/api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @swapkit/api
 
+## 3.4.3
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.3
+
+## 3.4.2
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.2
+
 ## 3.4.1
 
 ### Patch Changes