+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @coinmasters/wallet-evm-extensions
 
+## 8.0.3
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.3
+  - @coinmasters/toolbox-evm@8.0.3
+
+## 8.0.2
+
+### Patch Changes
+
+- bump for maya
+
+- Updated dependencies []:
+  - @coinmasters/types@4.4.2
+  - @coinmasters/toolbox-evm@8.0.2
+
 ## 8.0.1
 
 ### Patch Changes